Maintain a window's tiling state when moving across workspaces.
Implemented using a small cache of "dead" windows. When a window is removed from a given workspace,
it is marked as dead. But if it's then added to another workspace immediately, it gets rescued
and marked as active again. This allows us to persist the "tile_preference" across workspace
movements.

Garbage collection of dead windows is performed after the overview closes, or when there are 20
dead windows.